+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changes in Matrix iOS SDK in 0.11.0 ()
 Improvements:
 * MXRestClient: Add Matrix filter API.
 * MXRoomState: Add a membersCount property to store members stats independently from MXRoomMember objects.
-* MXRoomSummary: Add a membersCount property to cache MXRoomState one
-* MXRoomSummary: Add a membership property to cache MXRoomState one
+* MXRoomSummary: Add a membersCount property to cache MXRoomState one.
+* MXRoomSummary: Add a membership property to cache MXRoomState one.
+* MXRoomSummary: add isConferenceUserRoom.
 * MXStore: Add Obj-C annotations.
 
 Bug fix: